# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "SalesERP API",
    "description": "Sales ERP Management System API",
    "version": "1.0.0"
  },
  "host": "erp-sales-backend.onrender.com",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "paths": {
...
    "/api/purchases/": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "search",
            "in": "query",
            "type": "string"
          },
          {
            "name": "paymentStatus",
            "in": "query",
            "type": "string"
          },
          {
            "name": "isActive",
            "in": "query",
            "type": "string"
          },
          {
            "name": "fromDate",
            "in": "query",
            "type": "string"
          },
          {
            "name": "toDate",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "vendorName": {
                  "example": "any"
                },
                "vendorContact": {
                  "example": "any"
                },
                "purchaseNo": {
                  "example": "any"
                },
                "date": {
                  "example": "any"
                },
                "gstin": {
                  "example": "any"
                },
                "items": {
                  "example": "any"
                },
                "subTotal": {
                  "example": "any"
                },
                "cgst": {
                  "example": "any"
                },
                "sgst": {
                  "example": "any"
                },
                "igst": {
                  "example": "any"
                },
                "tax": {
                  "example": "any"
                },
                "discount": {
                  "example": "any"
                },
                "total": {
                  "example": "any"
                },
                "paymentStatus": {
                  "example": "any"
                },
                "amountPaid": {
                  "example": "any"
                },
                "balanceAmount": {
                  "example": "any"
                },
                "paymentMode": {
                  "example": "any"
                },
                "notes": {
                  "example": "any"
                },
                "isActive": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/purchases/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "vendorName": {
                  "example": "any"
                },
                "vendorContact": {
                  "example": "any"
                },
                "date": {
                  "example": "any"
                },
                "gstin": {
                  "example": "any"
                },
                "items": {
                  "example": "any"
                },
                "subTotal": {
                  "example": "any"
                },
                "cgst": {
                  "example": "any"
                },
                "sgst": {
                  "example": "any"
                },
                "igst": {
                  "example": "any"
                },
                "tax": {
                  "example": "any"
                },
                "discount": {
                  "example": "any"
                },
                "total": {
                  "example": "any"
                },
                "paymentStatus": {
                  "example": "any"
                },
                "amountPaid": {
                  "example": "any"
                },
                "balanceAmount": {
                  "example": "any"
                },
                "paymentMode": {
                  "example": "any"
                },
                "notes": {
                  "example": "any"
                },
                "isActive": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/purchases/{id}/payment": {
      "patch": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "amountPaid": {
                  "example": "any"
                },
                "paymentMode": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/purchases/soft-delete/{id}": {
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/purchases/restore/{id}": {
      "patch": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/purchases/deleted/list": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
  }
}